requirements.txt
#
#### Requirements for huesped_backend ####
django==1.9.7
djangorestframework==3.3.3
djangorestframework-filters==0.8.0
markdown==2.6.6
django-filter==0.13.0
django-rest-swagger==0.3.7
django-extensions==1.6.7
django-audit-log==0.7.0
django-cors-headers==1.1.0
django-reversion==2.0.6
psycopg2==2.6.1
gunicorn==19.6.0
requests==2.11.1
